Moved logic in settings.php into separate settings_xxx.php scripts and added a collapsible settings menu in the header.
